feat: use fixed integer size

+++ src/aoc_cxx/day01.cpp
@@ -12,14 +12,14 @@ namespace fs = std::filesystem;
 RegisterCommand day01("day01", {
     { "day01.txt",              381699,   111605670},
     { "day01_part1_test1.txt",  514579,   241861950},
-  }, [](fs::path filename) -> std::tuple<size_t, size_t> {
+  }, [](fs::path filename) -> std::tuple<uint64_t, uint64_t> {
 
     std::ifstream infile(filename);
     if (!infile.is_open()) {
       throw std::runtime_error(fmt::format("File Not Found : {}", filename.string()));
     }
 
-    std::vector<size_t> numberList;
+    std::vector<uint64_t> numberList;
     std::string line;
     while (infile >> line) {
       numberList.push_back(std::stoi(line));
@@ -51,14 +51,14 @@ RegisterCommand day01("day01", {
     return {part1Result, part2Result};
 });
 
-RegisterCommand day01Test("day01,test", {}, [](fs::path filename) -> std::tuple<size_t, size_t> {
+RegisterCommand day01Test("day01,test", {}, [](fs::path filename) -> std::tuple<uint64_t, uint64_t> {
 
     std::ifstream infile(filename);
     if (!infile.is_open()) {
       throw std::runtime_error(fmt::format("File Not Found : {}", filename.string()));
     }
 
-    std::vector<size_t> numberList;
+    std::vector<uint64_t> numberList;
     std::string line;
     while (infile >> line) {
       numberList.push_back(std::stoi(line));
